Redesign translation interface and enhance language selection
This commit overhauls the UI for the translation interface, improving usability and aesthetics. It introduces a  searchable dropdown for language selection, enhances the layout with Bootstrap cards, and fixes the copy to clipboard functionality. The display still needs to be improved around language selection, which is worse than originally done, but which is also more user-friendly in both what you can select and how you can select it.